ActionMailer + Mailcatcher: criaçao e configuraçao admin_mailer.rb, ao alterar o administrador um email e enviado.

+++ app/controllers/backoffice/admins_controller.rb
@@ -1,13 +1,12 @@
 # frozen_string_literal: true
 
+# Main Backoffice
 class Backoffice::AdminsController < BackofficeController
   before_action :set_admin, only: %i[edit update destroy]
-  after_action :verify_authorized, only: %i[new destroy]
+  after_action :verify_authorized, only: :new
   after_action :verify_policy_scoped, only: :index
 
   def index
-    # @admins = Admin.all
-    # @admins = Admin.with_restricted_access
     @admins = policy_scope(Admin)
   end
 
@@ -15,8 +14,6 @@ class Backoffice::AdminsController < BackofficeController
     @admin = Admin.new
     authorize @admin
   end
-
-  def edit; end
 
   def create
     @admin = Admin.new(params_admin)
@@ -28,8 +25,16 @@ class Backoffice::AdminsController < BackofficeController
     end
   end
 
+  def edit; end
+
   def update
+    passwd = params[:admin][:password]
+    passwd_confirmation = params[:admin][:password_confirmation]
+
+    check_password_blank(passwd, passwd_confirmation)
+
     if @admin.update(params_admin)
+      AdminMailer.update_email(current_admin, @admin).deliver_now
       redirect_to backoffice_admins_path,
                   notice: "Sucesso ao alterar: Administrador (#{@admin.email})!"
     else
@@ -38,7 +43,7 @@ class Backoffice::AdminsController < BackofficeController
   end
 
   def destroy
-    auhotize @admin
+    authorize @admin
     admin_email = @admin.email
     if @admin.destroy
       redirect_to backoffice_admins_path,
@@ -55,11 +60,6 @@ class Backoffice::AdminsController < BackofficeController
   end
 
   def params_admin
-    passwd = params[:admin][:password]
-    passwd_confirmation = params[:admin][:password_confirmation]
-
-    check_password_blank(passwd, passwd_confirmation)
-
     params.require(:admin).permit(policy(@admin).permitted_attributes)
   end
